# uint256: optimize MulMod, MulModWithReciprocal