--- a/mod.go
+++ b/mod.go
@@ -330,8 +330,8 @@ func Reciprocal(m *Int) (mu [5]uint64) {
 
 // reduce4 computes the least non-negative residue of x modulo m
 //
-// requires a four-word modulus (m[3] > 1) and its inverse (mu)
-func reduce4(x [8]uint64, m *Int, mu [5]uint64) (z Int) {
+// requires a four-word modulus (m[3] != 0) and its inverse (mu)
+func (z *Int) reduce4(x *[8]uint64, m *Int, mu *[5]uint64) *Int {
 
 	// NB: Most variable names in the comments match the pseudocode for
 	// 	Barrett reduction in the Handbook of Applied Cryptography.