The api of `reduce4` is changed to reduce array copies.

1. The function `reduce4` doesn't change the values of its parameters `x`, `m`, and `mu`. So it's fine to pass by pointers.
2. In the case of `z == m` (They are alias), `m` is used before `z`, and `z` is only written in the last step. So this is fine too.

## Benchmark
```
goos: linux
goarch: amd64
pkg: github.com/holiman/uint256
cpu: AMD Ryzen 7 7735H with Radeon Graphics         
                          │     old     │                 new                 │
                          │   sec/op    │   sec/op     vs base                │
MulMod/small/uint256-16     15.84n ± 2%   15.63n ± 1%   -1.33% (p=0.049 n=10)
MulMod/mod64/uint256-16     29.44n ± 1%   29.09n ± 2%   -1.17% (p=0.030 n=10)
MulMod/mod128/uint256-16    51.92n ± 1%   52.62n ± 2%        ~ (p=0.183 n=10)
MulMod/mod192/uint256-16    66.39n ± 2%   65.42n ± 1%        ~ (p=0.086 n=10)
MulMod/mod256/uint256-16    87.02n ± 2%   77.26n ± 1%  -11.20% (p=0.000 n=10)
MulMod/mod256/uint256r-16   38.07n ± 2%   27.52n ± 2%  -27.70% (p=0.000 n=10)
geomean                     41.79n        38.64n        -7.53%
```