@holiman 2024-05-27
Doesn't repro on my machine though: 
```
goos: linux
goarch: amd64
pkg: github.com/holiman/uint256
cpu: 12th Gen Intel(R) Core(TM) i7-1270P
                         │ bench.mulmod.1 │             bench.mulmod.2              │
                         │     sec/op     │     sec/op      vs base                 │
MulMod/small/uint256-8     36.33n ± 32%     47.02n ± 28%          ~ (p=0.105 n=10)
MulMod/mod64/uint256-8     57.80n ± 15%     65.16n ± 41%          ~ (p=0.143 n=10)
MulMod/mod128/uint256-8    108.5n ±  9%     138.3n ± 45%    +27.47% (p=0.019 n=10)
MulMod/mod192/uint256-8    148.2n ± 18%     230.3n ± 40%          ~ (p=0.063 n=10)
MulMod/mod256/uint256-8    197.3n ± 66%     204.8n ± 53%          ~ (p=0.565 n=10)
MulMod/mod256/uint256r-8   87.19n ± 47%     80.01n ± 35%          ~ (p=0.280 n=10)
MulMod/small/big-8         84.76n ±   ∞ ¹   81.11n ±   ∞ ¹        ~ (p=0.229 n=3+4)
```