@AaronChen0 2024-05-27
You can consider spinning up a ubuntu vps on DigitalOcean or Vultr for testing and delete the machine after finishing testing.